Split up memory optimization parts
